@@ -2,6 +2,7 @@ package eu.siacs.conversations.entities;
 
 import android.content.Context;
 
+import java.io.Serializable;
 import java.util.List;
 import java.util.Locale;
 
@@ -21,7 +22,7 @@ public interface ListItem extends Comparable<ListItem>, AvatarService.Avatarable
 
     boolean getActive();
 
-    final class Tag {
+    final class Tag implements Serializable {
         private final String name;
         private final int color;
         private final int offline;
