@@ -67,12 +67,16 @@ public class MessageGenerator extends AbstractGenerator {
         }
         if (message.edited() && !message.isMessageDeleted()) {
             packet.addChild("replace", "urn:xmpp:message-correct:0").setAttribute("id", message.getEditedIdWireFormat());
-        } else if (message.isMessageDeleted()) {
+        }
+        else if (message.isMessageDeleted()) {
             Element apply = packet.addChild("apply-to", "urn:xmpp:fasten:0").setAttribute("id", (message.getRetractId() != null ? message.getRetractId() : (message.getRemoteMsgId() != null ? message.getRemoteMsgId() : (message.getEditedIdWireFormat() != null ? message.getEditedIdWireFormat() : message.getUuid()))));
             apply.addChild("retract", "urn:xmpp:message-retract:0");
             packet.addChild("fallback", "urn:xmpp:fallback:0");
             packet.addChild("store", "urn:xmpp:hints");
             packet.setBody("This person attempted to retract a previous message, but it's unsupported by your client.");
+        }
+        for (Element el : message.getPayloads()) {
+            packet.addChild(el);
         }
         return packet;
     }
@@ -141,7 +145,6 @@ public class MessageGenerator extends AbstractGenerator {
         }
     }
 
-
     public MessagePacket generateChat(Message message) {
         MessagePacket packet = preparePacket(message);
         if (message.hasFileOnRemoteHost()) {
@@ -162,10 +165,9 @@ public class MessageGenerator extends AbstractGenerator {
 
             packet.addChild("x", Namespace.OOB).addChild("url").setContent(fileParams.url);
         }
-        packet.setBody(message.getQuoteableBody());
+        if (message.getQuoteableBody() != null) packet.setBody(message.getQuoteableBody());
         return packet;
     }
-
 
     public MessagePacket generatePgpChat(Message message) {
         MessagePacket packet = preparePacket(message);
@@ -174,6 +176,8 @@ public class MessageGenerator extends AbstractGenerator {
             final String url = fileParams.url;
             packet.setBody(url);
             packet.addChild("x", Namespace.OOB).addChild("url").setContent(url);
+            packet.addChild("fallback", "urn:xmpp:fallback:0").setAttribute("for", Namespace.OOB)
+                    .addChild("body", "urn:xmpp:fallback:0");
         } else {
             if (Config.supportUnencrypted()) {
                 packet.setBody(PGP_FALLBACK_MESSAGE);
