@@ -169,7 +169,9 @@ public class RtpSessionActivity extends XmppActivity
         setSupportActionBar(binding.toolbar);
 
         binding.dialpad.setClickConsumer(tag -> {
-            requireRtpConnection().applyDtmfTone(tag);
+            final JingleRtpConnection connection =
+                    this.rtpConnectionReference != null ? this.rtpConnectionReference.get() : null;
+            if (connection != null) connection.applyDtmfTone(tag);
         });
 
         if (savedInstanceState != null) {
@@ -802,8 +804,9 @@ public class RtpSessionActivity extends XmppActivity
                             .getJingleConnectionManager()
                             .getTerminalSessionState(with, sessionId);
             if (terminatedRtpSession == null) {
-                throw new IllegalStateException(
-                        "failed to initialize activity with running rtp session. session not found");
+                Log.e(Config.LOGTAG, "failed to initialize activity with running rtp session. session not found");
+                finish();
+                return true;
             }
             initializeWithTerminatedSessionState(account, with, terminatedRtpSession);
             return true;
