@@ -35,7 +35,11 @@ public class PhoneNumberUtilWrapper {
     }
 
     public static String normalize(Context context, String input) throws IllegalArgumentException, NumberParseException {
-        final Phonenumber.PhoneNumber number = getInstance(context).parse(input, LocationProvider.getUserCountry(context));
+        return normalize(context, input, false);
+    }
+
+    public static String normalize(Context context, String input, boolean preferNetwork) throws IllegalArgumentException, NumberParseException {
+        final Phonenumber.PhoneNumber number = getInstance(context).parse(input, LocationProvider.getUserCountry(context, preferNetwork));
         if (!getInstance(context).isValidNumber(number)) {
             throw new IllegalArgumentException(String.format("%s is not a valid phone number", input));
         }
