@@ -1,13 +1,15 @@
-package eu.siacs.conversations.xmpp;
+package eu.siacs.conversations.xmpp.forms;
 
+import com.caverock.androidsvg.SVG;
+import com.caverock.androidsvg.SVGParseException;
 import java.util.ArrayList;
 import java.util.List;
-
 import eu.siacs.conversations.xml.Element;
 
 public class Option {
     protected final String value;
     protected final String label;
+    protected final SVG icon;
 
     public static List<Option> forField(Element field) {
         List<Option> options = new ArrayList<>();
@@ -20,12 +22,21 @@ public class Option {
     }
 
     public Option(final Element option) {
-        this(option.findChildContent("value", "jabber:x:data"), option.getAttribute("label"));
+        this(
+                option.findChildContent("value", "jabber:x:data"),
+                option.getAttribute("label"),
+                parseSVG(option.findChild("svg", "http://www.w3.org/2000/svg"))
+        );
     }
 
     public Option(final String value, final String label) {
+        this(value, label, null);
+    }
+
+    public Option(final String value, final String label, final SVG icon) {
         this.value = value;
         this.label = label == null ? value : label;
+        this.icon = icon;
     }
 
     public boolean equals(Object o) {
@@ -39,4 +50,15 @@ public class Option {
     public String toString() { return label; }
 
     public String getValue() { return value; }
+
+    public SVG getIcon() { return icon; }
+
+    private static SVG parseSVG(final Element svg) {
+        if (svg == null) return null;
+        try {
+            return SVG.getFromString(svg.toString());
+        } catch (final SVGParseException e) {
+            return null;
+        }
+    }
 }
