@@ -44,6 +44,7 @@ import eu.siacs.conversations.services.AbstractConnectionManager;
 import eu.siacs.conversations.utils.CryptoHelper;
 import eu.siacs.conversations.xml.Namespace;
 import eu.siacs.conversations.xml.Element;
+import eu.siacs.conversations.services.XmppConnectionService;
 import eu.siacs.conversations.xmpp.Jid;
 import eu.siacs.conversations.xmpp.OnIqPacketReceived;
 import eu.siacs.conversations.xmpp.jingle.stanzas.Content;
@@ -137,9 +138,16 @@ public class JingleFileTransferConnection extends AbstractJingleConnection imple
                 try {
                     xmppConnectionService.getFileBackend().setupRelativeFilePath(message, new FileInputStream(file), extension);
                     finalFile = xmppConnectionService.getFileBackend().getFile(message);
-                    file.renameTo(finalFile);
+                    boolean didRename = file.renameTo(finalFile);
+                    if (!didRename) throw new IOException("rename failed");
                 } catch (final IOException e) {
                     finalFile = file;
+                    message.setRelativeFilePath(finalFile.getAbsolutePath());
+                } catch (final XmppConnectionService.BlockedMediaException e) {
+                    finalFile = file;
+                    file.delete();
+                    message.setRelativeFilePath(null);
+                    message.setDeleted(true);
                 }
 
                 xmppConnectionService.getFileBackend().updateFileParams(message, null, false);
